Integrate Spring Boot into PostManager class
Added Spring Boot annotations to PostManager class and commented out the load() method. The EntityManager is now injected through Spring's dependency injection mechanism, removing the need for the previous hard-coded instantiation. The load() method, which was manually loading entities from the database, is now not used as requirements have changed.