Update UI navigation and adjust user setting logic
Modified the Explore UI by updating the navigation components on the screen and adjusted the way users are being set in InstagramProfile UI. Specifically, the back button was moved to the top panel and the action taken when a username is clicked was tweaked. For the InstagramProfile UI, a temporary user was introduced to handle cases when the current user isn't set yet.