Change setting user method in SearchUI and ExploreUI
The method to set the user in the profile view of the application has been changed. In both the SearchUI and the ExploreUI, instead of setting the current user directly when a user panel or the username button is clicked, it now sets a temporary user. This change helps facilitate user interaction within the application, allowing for more flexibility when switching profiles.