Update SQL queries and enhance database connection handling
The existing SQL truncate query has been simplified in the EntityManager class. A new "dynamicQuery" method has also been introduced for executing complex SQL statements. Connection handling in the DatabaseConfig class has been improved with the use of Lombok's Setter.

In addition, created a new "queries.properties" file and a new "QueryConfig" class to manage SQL queries for the application. Multiple views in the "views.sql" file have been consolidated for clarity, and the DatabaseTest class was refactored to DatabaseTestEntity.